Extract `strings/combine_tests.cpp` into 3 separate cpp files
